UpiPlanet/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
httpx
prometheus-fastapi-instrumentator
sentry-sdk[fastapi]==2.44.0
pytest
redis
arq
inquirer
PyJWT
cryptography
tenacity
sqlalchemy
asyncpg
alembic
pgvector
bcrypt
